admin/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = ""
authors = [
    {name = "Neftalí Taiel Toledo Dicroce",email = "toledoneftali@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    # Flask
    "flask>=3.1.2,<4.0.0", # Framework web
    "flask-sqlalchemy>=3.1.1,<4.0.0", # ORM
    "flask-session>=0.8.0,<0.9.0", # Para manejo de sesiones

    # Base de datos
    "psycopg2-binary>=2.9.10,<3.0.0", # Driver para postgresql
    "geoalchemy2>=0.15.0,<1.0.0", # Compatibilidad para datos geoespaciales

    # Para los schemas (etapa 2)
    # "marshmallow>=4.0.1,<5.0.0",

    # Seguridad y validación
    "bcrypt>=4.2.1,<5.0.0", # Hasheo de contraseñas
    "flask-bcrypt (>=1.0.1,<2.0.0)", # Integración de bcrypt con flask
    "flask-wtf>=1.2.2,<2.0.0", # Compatibilidad de flask con wtforms
    "wtforms>=3.2.1,<4.0.0", # Formularios y validación de datos
    "python-dateutil>=2.8.2,<3.0.0", # Manejo de fechas
    "email-validator>=2.3.0,<3.0.0", # Requerido para validar emails en wtforms

    # Configuración de entorno
    "python-dotenv>=1.0.1,<2.0.0", # Permite cargar el .env para configurar el entorno

    # Utilidades
    "python-slugify>=8.0.4,<9.0.0", # Para crear slugs fácilmente
    # "pillow>=10.4.0,<11.0.0", # Para manejo de imágenes (probablemente sea útil en la etapa 2, lo dejo por las dudas)
    "folium>=0.17.0,<1.0.0", # Para los mapas interactivos
    "pandas>=2.1.0,<3.0.0", # Para los CSV y exportación

    # Para un logging más legible
    # "python-json-logger>=2.0.7,<3.0.0",

    # Necesitemos esta librería para que geoalchemy funcione mejor y es requerida para la segunda etapa
    "shapely (>=2.1.2,<3.0.0)",

    # Para permitir acceso desde el portal público
    "flask-cors (>=6.0.1,<7.0.0)",
    "minio (>=7.2.18,<8.0.0)",
]

[tool.poetry]
packages = [{include = "web", from = "src"}]


[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"
pytest-flask = "^1.3.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
flask-shell-ipython = "^0.5.3"
flask-debugtoolbar = "^0.15.1"
faker = "^33.1.0"
autopep8 = "^2.3.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"